+++ OneBlink.SDK/models/FormElementLookupPayload.cs
@@ -1,0 +1,25 @@
+
+using Newtonsoft.Json;
+
+namespace OneBlink.SDK.Model
+{
+	public class FormElementLookupPayload<T>
+	{
+		[JsonProperty]
+		public Form definition
+		{
+			get; set;
+		}
+		[JsonProperty]
+		public FormElement element
+		{
+			get; set;
+		}
+
+		[JsonProperty]
+		public T submission
+		{
+			get; set;
+		}
+	}
+}